changed how the tcp server is created
