+++ src/ToolKit/Communication/TcpMessageReceived.cs
@@ -1,0 +1,3 @@
+namespace FatCat.Toolkit.Communication;
+
+public delegate void TcpMessageReceived(byte[] data);
